requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asgiref==3.2.10
astroid==2.3.3
asttokens==2.0.4
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
dialite==0.5.2
distlib==0.3.0
Django==3.1.1
docutils==0.16
filelock==3.0.12
flexx==0.8.0
idna==2.9
imagesize==1.2.0
ipython-genutils==0.2.0
isort==4.3.21
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
markdown2==2.3.9
MarkupSafe==1.1.1
mccabe==0.6.1
meta==1.0.2
nbformat==5.0.6
packaging==20.4
pathspec==0.8.0
Pillow==7.2.0
pscript==0.7.4
pycodestyle==2.6.0
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQtWebEngine==5.15.0
pyrsistent==0.16.0
pyshortcuts==1.7.1
pytz==2020.1
pywin32==227
regex==2020.5.14
requests==2.23.0
setupext-janitor==1.1.2
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.20
webruntime==0.5.7
windows-curses==2.1.0
wrapt==1.11.2
